feat: Refatorando tela de roteiro para mvc
