+++ src/pages/roteiro/RoteiroStyles.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const Root = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
+
+export const ToolbarDiv = styled.div`
+  width: 100%;
+`;
+
+export const SpinDiv = styled.div`
+  height: calc(100vh - 80px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
